services/command-service/requirements.txt
kafka-python==2.2.15
fastapi==0.120.4
uvicorn[standard]==0.38.0
pydantic==2.12.4
